pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.0",
    "google-generativeai==0.8.3",
    "httpx==0.27.2",
    "numpy==2.1.3",
    "python-dotenv==1.0.1",
    "soundfile==0.12.1",
    "uvicorn[standard]==0.32.0",
    "vadersentiment==3.3.2",
    "vosk==0.3.45",
]
